% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixout_conti.R
\name{fixout_conti}
\alias{fixout_conti}
\title{fixout_conti}
\usage{
fixout_conti(
  dt,
  col_name,
  interactive = FALSE,
  iden.m = "resistant",
  fix.m = "asNABoth",
  rangeLU = NULL,
  k = 1.5,
  exclude = NA,
  logt = FALSE,
  plot = FALSE
)
}
\arguments{
\item{dt}{A data frame or a data.table.}

\item{col_name}{A string indicates column name with data need to be fixed.}

\item{interactive}{TRUE/FALSE, whether to fix the outliers in an interactive way.}

\item{iden.m}{A string shows the chosen method for identifying outliers."resistant": standard boxplot; "asymmetric": modification of standard method to deal with (moderately) skewed data;"adjbox": adjusted boxplot for skewed distributions;"winsorize": the smallest and/or the largest values are replaced by less extreme values.}

\item{fix.m}{A string shows the chosen method for fixing outliers. "dropBoth": drop the whole row where outliers are from both sides of the data; "dropLeft": drop the whole row where outliers are from left side of the data; "dropRight": drop the whole row where outliers are from right side of the data; "asNABoth": convert outliers from both sides of the data as missing values; "asNALeft": convert outliers from left side of the data as missing values,"asNARight": convert outliers from right side of the data as missing values; "median": replace outliers with median; "mean": replace outliers with mean.}

\item{rangeLU}{The lower and upper limit of range for the data.}

\item{k}{A constant to determine the lines outside the upper and lower quartiles.}

\item{exclude}{Values that will be excluded in the numbers to be processed. Missing values are removed by default when detecting outliers.}

\item{logt}{TRUE/FALSE, whether the numbers are transformed with an lognormal distribution.}

\item{plot}{TRUE/FALSE, whether plots are shown.}
}
\value{
A data table contains the fixed column.
}
\description{
This utility identifies and replaces outliers for continuous variables. It assumes that the columns containing continuous variable is a data.frame and replaces the column elements that are outlier with NAs. It also provide the density plot for before and after data for users to make assessment on the quality of the data post outlier removal. Users have an option to drop the rows with outlier or keep them after replacement. By default the function will retain the rows. Optionally you can replace these values with their mean or median.
}
\examples{
data <- as.data.table(mtcars)
output_table <- fixout_conti(data,col_name="wt",rangeLU=c(2.5,3.5))
}
